Re-organize integration test modules to be able to run O_DIRECT tests serially
