+++ mountpoint-s3/Cargo.toml
@@ -60,6 +60,7 @@ proptest = "1.0.0"
 proptest-derive = "0.3.0"
 rand = "0.8.5"
 rand_chacha = "0.3.1"
+serial_test = "2.0.0"
 sha2 = "0.10.6"
 shuttle = { version = "0.5.0" }
 tempfile = "3.4.0"
